RealtimeEffectList guards changes with a spin lock...
... Main thread constructs a replacement array of states, and holds the spin
lock only for the duration of a vector swap to make the changes.

Worker thread never changes the contents of the list, but must lock for the
complete duration of calls to ProcessStart, Process, ProcessEnd of the
RealtimeEffectManager, to ensure exactly the same states are visited starting
and ending and processing in-between.

So the main thread might possibly spin for the whole duration of the realtime
processing of a block of samples.  But it is more important to have the
worker rarely block and use a light weight locking protocol.

One modification of RealtimeEffectLists::mStates does not need a lock guard:
it is for deserialization only.