Issue #2909: Export Multiple dialog: focus issue on Windows
Problem:
When the Export Multiple dialog opens on Windows, the first control is not the focus.

This bug was introduced between Audacity 2.4.1 and 2.4.2. During this period there were no significant changes to the code of this dialog. However the version of wxWidgets changed from 3.1.1 to 3.1.3, so this is probably a wxWidgets bug.

Fix:
Explicitly setting the first control to be the focus, as a work around for the assumed wxWidgets bug.