Implement activation per RealtimeEffectList...
... which does not forget the individual states' activation when you turn it
off, but does temporarily override any "on" states.

Changing of the value is mutually exclusive with RealtimeProcessingScope, so
that the activation state for each list is also unchanging in that scope.

The meaning of the boolean passed to state-visiting functions is changed.  If
the state's own activity state is needed, the state itself can tell.

Note that this also leaves RealtimeEffectState::IsActive called only in the
processing thread.