RealtimeEffectState holds two copies of EffectSettings...
... One, immediately updated by the main thread, and used for pushing undo
states and persistency.

The other, getting delayed but atomic updates via the EffectSettingsAccess
subclass while doing realtime processing.  It is not copied to the undo stack.
It is also copied to the main thread, as the result of
EffectSettingsAccess::Get().

This avoids data races.

Still to do, to make settings "stick", is to copy from the first to the second,
always (whether processing realtime effects in another thread or not), without
races, and quickly enough in response to the user interface changes.

The two copies are always synched when the main thread first calls GetAccess.