Issue #2912: incorrect location of focused track for AT
Problem:
Location of focused track provided for assistive technology is incorrect - it should be roughly the same as the yellow outline.

In TrackPanelAx::GetLocation(), in the call rect.SetPosition(), in the case of a focused track the Track Panel's parent (the Main Panel) is incorrectly being used. This didn't cause a problem as the top left corner of the Track Panel and the Main panel used to be the same. However, during 3.2.0 development, a change was made so that the Main panel now includes the Timeline, and so the top left hand corners no longer coincide.

Fix:
For a focused track, don't use the Track Panel's parent in the call rect.SetPosition()