issue i2911: wrong duration in built-in Generators (Turkish)
When Audacity is set to use Turkish, the built-in "Generate" effects fail to set their default duration to match the selection.

Caused by typo in tr.po
Note: This after this fix the duration is set correctly. However, following the steps to reproduce, there is a problem with the length of the resulting first and second clips being the wrong length. This is a separate bug and occurs when the language is set to English (and I presume others).