MakeInstance takes no argument...
... Not ever loading the current settings group; instead, assume that was done
to the EffectSettings object already if necessary.

That will not be necessary -- it would be wrong -- where the RealtimeEffectState
uses MakeInstance() with independent settings.

It is right in the other uses of MakeInstance -- for applying an effect
destructively, or for getting default macro parameters just for a macro
editing dialog.

And the current settings group is persisted whenever destructive effect
application is done, and then reloaded again in InitializePreset() in
EffectManager.cpp, the first time any effect is used again in a session, which
initializes what GetDefaultSettings() will return.