Fix initial discovery of Ladspa and AudioUnits effects...
... Problem was GetConfig() tried to use the configuration file in the
PluginHost process, but PluginManager::Initialize had not been done, and the
result was a std::bad_function_call exception.

Solution is to delay the reading of configuration file values until the effect
is really used.  The values aren't needed just for plug-in discovery.

LV2, VST, and VST3 code did not contain any similar early uses of GetConfig.