Single function to format effect names in RealtimeEffectPanel...
... not making the assumption that a non-null effect pointer is available for
the stored effect ID.

This completes null-checking on the uses of RealtimeEffectState::GetEffect().

This parses the effect name from the ID; the system for persistent effect
identification in fact already included the name.

Only an un-translated name will be found in this case -- but this is not likely
to make a difference, because third-party effects probably won't have an
internationalized name available anyway, and built-in effects won't be missing
(unless, perhaps, the project file comes from a future version of Audacity with
more built-in effects).