Tricks to serialize and deserialize the new enum to config file...
... Write the string "00" which earlier Audacity will interpret as boolean
false -- some stepping through wxWidgets code confirms that this is the
implementation of the override of pRegistry->Read for a boolean value: string
to long, then long is expected to be zero or one.

But Audacity 3.2 reads a string value, not converted to bool, and con give a
different special interpretation to a leading 0 digit.