Recheck built-in effects for realtime capability at startup...
... but only when we detect that the config file needs to migrate.

This code is reached before PluginManager::Save().

This will update the cached information in the config file, in 3.2 or a future
version of Audacity, whether upgrading the config file to a new version, or
downgrading to a previous version of Audacity, (with also a different plugin
registry version) that is at least Audacity 3.2.

It will NOT solve the case that someone downgrades from 3.2 back to 3.1.  Then,
in 3.1, some effects (Reverb at least) will be mistreated as realtime with a
non-modal dialog.  They might still be applied.  In this unlikely case, support
can direct a user to delete or edit pluginregistry.cfg.

Every type of built-in effect is constructed once at startup time when upgrade
or down is needed, but then this is skipped on later restarts, because the
plugin registry acts as a cache.

Precisely because it is a cache -- it had a cache invalidation problem, which
this PR fixes.