Adds verbose logging for Journal tests on CI
It was problematic to determine the reason why the sanity test has failed,
especially on Windows. 

CTest framework does not allow to get the exit code of the application (or there is no easy way to retrieve it at lease).

To simplify the test run analysis the following was added to the journaling system:

* Journal writes every step and sync exception into `journallog.txt` with aggressive flushing
* On Windows - test run is wrapped into a PowerShell script, that outputs the exit code and collects the logs of the test run. It is unclear, if a similar script is needed on *nix systems