Rewrite fetch, store, load, save of parameters...
... Fetch stores complete information -- names as well and values, and whether
the parameter was available at all in the AudioUnit's state.  It doesn't clear
the map but reassigns nullopt for an unavailable parameter.

Store reassigns to all parameter IDs in the AudioUnit that did not have nullopt
in the map.  Store requires two passes, as explained in comments.

Save iterates the map only, not using the AudioUnit handle.

Load iterates the config file and updates every member of the map to nullopt or
a value from the file, also not using a handle.