issue 3039: vst3 plain gui: accessibility of sliders
Screen readers do not automatically read the the actual values of sliders which is shown in the text to the right of each slider.

Fix:
Set the accessibility name of the slider to be the name of the control followed by the actual value of the slider. This is the same as was done for the vst plain gui.