issue #3021: vst3 plain gui: accessibility of check boxes
Problem:
The mouse interaction and focus of check boxes in the vst3 plain gui is non-standard, and the names of the check boxes are not read by screen readers.

This is caused by creating a wxCheckBox with an empty label, and adding the text separately.

Fix:
Use a wxCheckBox with a label, and with the wxALIGN_RIGHT style so that the label is to the left of the check box which is the desired layout in this case.