Fix deadlock adjusting effect dialog during pause...
... Lifting the ProcessingScope object to the beginning of FillPlayBuffers
causes RealtimeEffectState::ProcessStart and RealtimeEffectState::ProcessEnd
to be visited in its ctor and dtor, even when there are no samples to consume
from the ring buffers in paused state.

That is in the audio thread. That makes some progress copying settings changes,
so that RealtimeEffectState::Access::Get in the main thread can exit the loop,
when it is expecting an "echo" of its last write.