Fix the infinite recursion when explicitly copying EffectSettings...
... Which happened on some compilers.

It seems they differ in rules for overloading resolution, of the generated
constructors of EffectSettings, and the inherited template generated
constructors of the base class TypedAny.

This fix in TypedAny uses sfinae to eliminate its constructor from the
candidate set of overloads when there is only one argument, and it is
lvalue- or rvalue- or non-reference, to const or non-const, type Tag.