Lower the responsibility to make the instance into dialog factory...
... So if the dialog is connected to a realtime effect state, it can hold a
shared pointer to that state's instance; and each time play starts, the state
will re-use that instance rather than making a new one.

This means: if you open the old non-modal dialog for an effect, then you can
play, stop, and play again (and again and...), and whenever you play you can
adjust sliders with the correct results.

The sidebar dialogs are not yet non-modal but this will be a requirement for
them too when they are.

Connecting the effect dialog to the same instance that is doing the realtime
processing will also be necessary to fix output metering of effects.

Note that some steps of InitializeIntance() should not be done for the sidebar
dialog, and the constructor knows the difference because a state is given to
the constructor.  Likewise in CleanupRealtime().