Setting will be marked as valid after calling Write inside the scope
Currently, there is a following problem with `Setting::Write` inside the scopes:

Consider a setting, that is not yet written to the config or has a default value in it.

Calling `Read`/`ReadWithDefault` on it will never mark the Setting as valid, because of the following condition:

`this->mValid = (this->mCurrentValue != defaultValue);`

That means, that every read for such `Setting` is satisfied through the config.

On the other hand, `Write` only marks Setting as valid if the `Write` is called outside the `SettingScope`. Write never checks, if the new value is different from the old value.

Given all that, it is impossible to `Read` value changed by `Write` inside the scope if originally the value was not changed to something not default.

This breaks the FFmpeg locating logic: we write a candidate directory to `AVFormatPath` inside a scope, and the subsequent read always returns the default value. 

The is fixed by marking Setting as valid inside any scoped write. `mValid` is not rolled back, but that makes no observable difference, because default value will be set to `mCurrentValue`