For AU, first RealtimeAddProcessor reuses this, doesn't allocate...
... This is important to connect the dialog to the right instance, for
output monitoring and for proper response to slider changes.

When using an AudioUnit effect in the per-track stack,
there is never another processor.

There may be multiple processors only in the legacy UI for
per-project effect stacks, when an AudioUnit applies to more than
one track at once.

Reviewed all uses of mSlaves; what is done to them in realtime
processing must also be done to self now.