Define and use EffectSettingsAccess::Flush() ...
... By calling it for the effect power button, this completes the fix for
persistency of effect power buttons.

By calling it when transferring data from the EffectUIHost window, we have
sufficient flushes to make up for the lost side-effects of Get() for the
realtime effect state accessor.

Persistency of the power button could have been fixed just by calling Get()
after ModifySettings().

However this more elaborate fix will also be useful for performance of effect
output metering (such as for AUDynamicsProcessor).  It used to be that each
Get() in ModifySettings() made a complete flush, but that sometimes made meter
updates and responses to slider movement very jerky, when the effect was opened
from the sidebar.