Eliminate SetSampleRate method of effect instances...
... Previous commits pass the value instead in arguments where it is needed.

The member variable StatefulEffectBase::mSampleRate is gone.

A few effects may store the value in ProcessInitialize for reuse in ProcessBlock
which is always called only after ProcessInitialize.

If a SetSampleRate() override had some important side effect, do that directly
in ProcessInitialize and RealtimeInitialize.  (Or just in the first, when the
second calls the first.)