Finally!  What motivated this long digression...
... Ladspa effects do not actually support changeable sample rates in instances:
see in ladspa.h how you must pass a sample rate when constructing one, and there
is no interface given for changing it afterward.

lv2 effects may support a settable sample rate as an option, and we had
code to discover whether each one does, but they don't necessarily.

So were our SetSampleRate overrides actually wrong, because they should
have been destroying and recreating the ladspa and lv2 instance objects in case
of immutable sample rates?

Maybe so, but now we have transformed code so it doesn't matter.  We don't
need SetSampleRate.

And therefore we don't need the complications of discovery of support of
mutable sample rate, so simplify lv2.

However I don't know whether any effects might require the host to support the
sample rate option.  So, the mSampleRate member of LV2FeaturesList is kept and
still passed in the array of options, and correctly reassigned before each
instance is created, in case it matters.

I used `mutable` to keep the features list object const.  To be fixed later
maybe when the features list is replicated in each of our instance objects.
Where features and options capture addresses of state variables, or of values
that may differ across instances -- then the really should be stored separately
for each instance, not in the effect object itself.