Non-inheritance among LV2Effect, LV2UIFeaturesList, LV2Features...
... And the basic list of features in the effect is const, though still with one
mutable member for sample rate.

Note that a shorter list of features is given to LV2Wrapper creation.
Presumably that will always still satisfy the requirements.

Note ValidateFeatures() is also called in BuildFancy() and there, the longer
list of features is validated against another "subject" uinode obtained first
by a query of the plug-in for its UI support.