issue #3013: play indicator not displaying correctly.
This is a partial fix for issue 3013, which also includes the scrub ruler not displaying correctly.

Problem:
Incorrect code which only worked when the parent of the track panel had the same top left hand corner as the track panel. Now that the timeline has been reparented, this is no longer the case.