issue #3013: Scrub ruler not displaying correctly
This is a partial fix for issue 3013, which also includes the play indicator not displaying correctly.

Fix:
Now that the timeline has been reparented by the main window, the size event only needs to be sent to the timeline's parent, not grandparent - no other windows are affected by the resize.