Block size bugs fixed; LV2InstanceFeaturesList separated...
... It holds all the options; some options (for sample rate and block size)
may need to vary between LV2Wrappers.

Commit 76028f62 introduced a subtle mistake, in that the lv2 plug-in was
pointed to the wrong location for block size, not the one inside the
LV2Instance that was really updated.

Now we correct that.  The features list, only, has a block size member.
Its initialization is delayed until the LV2Wrapper is made.

If multiple wrappers exist, be careful to replicate block size changes in all.

Be careful also, that LV2UIFeaturesList is always built now as an extension of
LV2InstanceFeaturesList.

Also don't need mutable for the sample rate any more, that's good.