Are we done NOW?  No...
... not until we stop passing the address of a temporary to the foreign lv2
code.

I'm not sure this could have caused harm, but I can't prove it won't with any
existing plug-in, so let's remove doubt.

The interface for features, unlike that for options, uses another level of
indirection (array of pointers to structs, not an array of structs) which makes
me think there is some assumption of permanence of the feature information,
while the instance exists, that we should respect.