TrackListEvent::DELETION behavior is changed
To be deleted track is passed in `mpTrack`,  `mExtra` reflects if this even is a part of track replacement or not.

Previously, none of the fields were used so this change should bring no observable difference.  It is worth noting though that Clear now fires an event for each track.