MPG123 importer now only allows to MPG123_NEW_FORMAT once
MP3 file is a number of the MPEG audio frames stitched together and interleaved with
any other data. This potentially allows to mix different output formats inside a single file.

This, however, is not handled by Audacity import code at all.  libmpg123 has `MPG123_NO_FRANKENSTEIN` which disallows changing the format between the frames. 
It was observed that `MPG123_NO_FRANKENSTEIN` results in successful imports of files, that
are not MP3 files. For this reason importer now allows `MPG123_NEW_FORMAT` only once. The respective code is moved to `Open`, as it allows to bail out earlier for the files that are not MP3s.