In case of failed import allow other import plugins to run
It is not always possible to guess the file type on `Open`, without trying to import it. This is especially
true for MP3 files, where only an attempt to import the file can give a somewhat definitive answer about
the file format. Several files were observed, where libmpg123 detected mp3 like structures, but that were
not MP3 files.