Forward OnClose event of the EffectHostUI to the Effect
As a part of transition to the new Effect architecture `EffectUIValidator` subclasses now handle the events from wxWidgets. This makes deleting `EffectUIValidator` instance from the `EffectUIHost::OnClose` to be unsafe and in fact it has introduced #3220: LV2Validator called `Close` from the `OnIdle` handler, which deleted the handler mid-processing.  

However, the destructor of `EffectUIValidator` had a side effect of calling `CloseUI` on the plugin itself. This broke Nyquist debug mode: after pressing Debug button `EffectUIHost::Destroy` was called, which has delayed the `EffectUIHost` and `EffectUIValidator` destructors until the next idle event in the run loop. However, `NyquistEffect` relied on `CloseUI` to correctly setup parent of the `NyquistOutputDialog`. Omission of `CloseUI` from the `EffectUIHost::OnClose` results in attempt to `delete` a stack object, because `NyquistOutputDialog` is attached to the `EffectUIHost` in the destroying state.