Fix destructive application of LV2 effects...
... Because the contract of EffectInstance does not forbid setting of block
size before process initialization, we accommodate it in LV2Instance by
separating the Instance features list that holds options, from the Wrapper
features list that holds a feature needed for each lv2 instantiation.

This allows setting of block size even when the lv2 instantiation does not yet
exist.

The Options are passed to lv2 by const void *, and the locations can be shared
among master and slaves, but the LV2_Worker_Schedule object needs to hold
a different `this` in each of them.