More member initializers and const members for Mixer...
... And two members made constexpr.

mBufferSize was not used after construction but is now returned by an accessor,
and is needed to state a postcondition of construction, and precondition for
Process().

Also added some assertions in private functions supporting proof of the post
of Process().

In fact, no uses of Mixer now pass less than the full buffer size, except in
AudioIO.cpp.

Please verify sequence of initializers is correct, and none depends on later
members.