RealtimeEffectManager::Process is given enough buffers...
... So the extra "dummy" buffer isn't needed to redirect extra output.

And extra input buffers are initialized with zeroes, and passed to the effect
processor, instead of replicating mono channels.

This gives realtime play the same inputs as would be given to destructive
effect application, when the track is mono.  This makes a difference for
reverb effects, in which each output channel is influenced by both of the input
channels.  Channels are not processed by reverbs in isolation.