Adds a class to manage stacked effects settings UI
Whenever track data is modified (by destructive effect or in the end of recording) - Audacity replaces the track with a cloned one.

Cloning performs deep copy of RealtimeEffectStateList attached to the track. This copy involves copying
of the weak pointer to effect instance, created by the respective RealtimeEffectState (if any).

Sharing `mwInstance` is considered to be safe:

* If the effect is used inside as a real-time processor - the instance cannot change until the 
   processing is complete.
* `RealtimeEffectState` is not used for destructive processing
* If UI is shown - it is expected that the instance held by the dialog can be used for the subsequent 
   real-time processing. Otherwise it breaks the current "destructive" flow. While that flow may change
   in the future we can rely on that behavior now.

The last case requires `EffectInstance` to be independent of the sample rate used. Otherwise it is not possible to open a dialog that survives the start of the playback, because the current implementation applies the effect AFTER the resampling. `RealtimeEffectStateUI` will close the UI if it detects the instance mismatch.