Fix free store allocation that was hiding in EffectSettingsAccess...
... This is the simplest fix to write for now.

Note that the get and set of duration format happen only in the main thread, so
no harm in simply sharing the state.

In the next version we should rewrite the inter-thread communication more
thoroughly and separate concerns better, so that the structure holding
persistent settings may be different from those carrying messages between
threads.