LabelTrackView: Horizontal label text layout
This replaces the logic for horizontal label positioning with a
piecewise continuous function that successfully centers label text while
both label endpoints are within the viewport, while still retaining the
secondary features of label positioning: Hinting at position within the
label when the endpoints are offscreen, and allowing the user to scroll
through the text when it is too wide to fit in the viewport.

It is also easier to adjust the behavior of this by choosing another
sigmoid function to control how the off-screen handles translate into
onscreen motion.