Fix the assignment of processors to channels...
... The number of output buffers to fill up depends on hardware, not the track
or the effect.  This will let effects like BC Chorus Mono play on a stereo
track without producing junk in the right speaker.

This commit is all about passing that number to RealtimeEffectState::AddTrack()
and passing the same later to RealtimeEffectState::Process() which both use
AllocateChannelsToProcessors().

This supersedes the quick fix at 2f60f82, so issue 3443 should be verified
again.