LabelTrackView: Dynamically color glyphs
Split label glyphs into two layers, one which is multiplied by the
track base color, and one which is overlaid on top with fixed (theme)
color elements. This paves the way for label tracks to take on
arbitrary colors while still allowing the glyphs to be themed.

There is one visual change: The "dark" and "high-contrast" themes used
colors that were slightly different than the label track colors and I
did not attempt to author the masks to recreate those colors.