Lift EffectContext construction out of EffectUI::DoEffect...
... and out of dialog construction, into many other places.

This requires a guarantee of the lifetime of the context object in case dialogs
are non-modal, which is given by shared_ptrs to context stored in the dialogs.

These many places are invocations of effect application as in applying macros,
or doing the "Repeat last effect" command, and similar.

It also requires a context when the sidebar opens a dialog.  But the effects
available there are PerTrackEffects which should not be context dependent for
processing.

This lets preview construct a temporary, modifiable context from that given
to dialog construction:  it will not need to invent completely new context where
it otherwise lacks the information to do that.

We see here another reason that ShowHostInterface ends up needing context
anyway, not just because of the peculiarities of NoiseReduction.

EffectUI::DoEffect no longer takes a CommandContext object.