Revert parts of 378ae8d and 652588d...
... The data transfers to and from effect dialogs, for built-in effects (or
Vamp, or Nyquist) which are not yet part of the stateless system, were not in
all cases made superfluous, as seen from some known issues including 3353,
2762, 3234

So just put them back into all of them.  For other effects this may fix bugs we
haven't found or else it is harmlessly redundant.