Issue 3565: GVerb crashes in effect stack of stereo track...
... This simple fix corrects the crash.  Whether the resulting output is correct
is a matter of definition.

GVerb takes one channel in and always writes two channels out.  When rendering
the right channel then, supply an extra throw-away buffer.

What happens in the result is that the left output, only, of each channel is
retained.

What really should be done?  It might make sense to mix the two pairs of output
channels down to one pair, losing none of the output.

That is outside the scope of this simple fix.