Issue 3792: Various play commmands do not exit play mode
This affects "play cut preview", and most of the play before/after commands on the transport sub menu of the extra menu.

Whether or not the commands exit play mode depends in some apparently random way on one or more values of selection start, selection end, and in the playback page of preferences: the values of cut preview before and after cut region.

Play cut preview:
In CutPreviewPlaybackPolicy::GetPlaybackSlice(), in the initialization of Samples2, the fractional part of mDuration2 * mRate is disregarded. This can result in CutPreviewPlaybackPolicy::Done() never returning true.
The fix is to round the value of the float, rather than disregarding the fractional part.
The initialization of Samples1 was also changed, just for consistency.

Play after/before commands.
A similar issue.
In PlaybackPolicy::GetPlaybackSlice(), assignment of toProduce was changed to include rounding, and the assignment of frames was changed for consistency.

Note that in NewDefaultPlaybackPolicy::GetPlaybackSlice(), rounding is already used:
toProduce = frames = 0.5 + (realTimeRemaining * mRate) / mLastPlaySpeed;