issue 3370: accessibility of toggle buttons in RT effects panel
Problem:
Narrator doesn't read the pressed state at all, and Jaws and NVDA don't read the toggle buttons as expectd.

Fix:
In MSAA, there isn't such a thing as a toggle button. So imitate one by including the role and the state in the name, and creating a name change event when the state changes.