i3911: Audio setup/share buttons should not be toggle buttons
The Audio share and Audio setup buttons are set to be toggle buttons, which they are not. This affects both sighted users and users of screen readers.

Fix:
The buttons are no longer set to be toggle buttons.
Note that with this fix the buttons are not shown as down at the point when they are clicked, which they ideally should be. However, the menu opening gives visual feedback that the button has been pressed, and there are a number of other "menu buttons" already in Audacity which have the same behaviour.