Do less in the AudioUnitValidator event handler...
... Update only the changed settings.  Do not fetch all of them.  Makes opening
the dialog faster, but not yet fast enough.

When some effects (like Waves Studio Rack) first open the dialog or change the
preset, they send messages for each parameter.  When there were many (2000)
parameters, the quadratic number of calls to fetch parameters was noticeable.

Eliminating FetchSettingsFromInstance is not enough yet, because there are still
many complete copies of AudioUnitEffectSettings done in the calls to
ModifySettings, so there is still some quadratic work.