Yet another hotfix for LWSlider's tooltip
Originally, crash was caused by misuse of `SetParent` - should be `Reparent` instead, but then another issue appear.
`LWSlider` gets re-parented inside DC drawing and hit test routines and parents seem to differ from each other. It's not clear what is the purpose of such re-parenting, but tooltip starts to flicker after that. So the hotfix is to omit tooltip re-parenting.