Fix data loss in crash recovery after recording...
... This data loss happened only when recording completed normally, but then
a crash happened before any other edit that pushed undo history.  And it
only lost the last 1 MB block or less in each channel, which is a bit less than
6s with default sample rate and format.

See commit 06f22e94, which disabled auto-save for the completion of recording,
though other auto-saves happen at the completion of each block during recording
in `OnAudioIONewBlocks`.  The good reason for that change was that there
may have been an exception detected during recording itself because of
exhaustion of disk space.  But in the usual case this won't be needed.