mUIParent is private in Effect...
... It is only ever assigned non-null in PopulateUI() as S.GetParent().

So let any effects that need it grab their own copies of the pointer in
PopulateOrExchange.

For effects that have been made stateless, not the effect, but the validator,
will hold the pointer.

This almost eliminates another piece of state of the effect object (though only
UI related) that "stateless" effects were still using.

Yet Effect::mUIParent is still used in the protected member function MessageBox
which will be rewritten later to eliminate it wholly.