Fix translation of "host" in Audio Setup tool bar
The german "Rechners" refers to a computer. All other instances of "host" or "audio host" are kept as "Host", so use it in Audio Setup tool bar as well.
Further the keyboard shortcut "A" was used twice, although I was not able to us keyboard navigation in Windows in that toolbar.