Issue #3482: accessibility of power toggle buttons in dialogs
Problem: The state of the power toggle buttons in realtime effects dialogs are not accessible for users of screen readers.

Fix: A wxBitmapButton was previously used for the power toggle button. However it's not possible to imitate a toggle button for all the Windows screen readers using this control.
So use an AButton instead. This is already used for the power buttons in the realtime effects panel, and has already been made to imitate a toggle button for all Windows screen readers.

Notes:
- Using an AButton means that it probably won't be easily possible to assign an access key to this button. But this is a minor issue compared to the state of the button not being accessible.
- The accessibility name of the toggle button was changed from Enable to Power to make it consistent with the power toggle buttons in the realtime effects panel.
- In EffectUIHost::InitializeInstance(), a line updating the state of the toggle button was removed, as it was immediately prior to a call to UpdateControls() which updates the state of the toggle button.