EffectBase (and others) -> ProjectAudioIO broken by a global hook...
... for default playback policy; ProjectAudioManager injects the full
implementation at startup.

ProjectAudioIO makes some use of lib-audio-graph directly (which seems
reasonable) but avoids using lib-screen-geometry even though it already has
that transitive dependency.

Rather, leave things related to the looping region of the ruler in the high
level injected function that extends the behavior of the default function.