[NYQUIST] fix nyq_reformat_aud_do_response() not found. (#4322)
If NYQUIST is disabled when configuring Audacity, the build process hangs with an error because nyq_reformat_aud_do_response() is not found when linking, see issue #3137. Since ExecForLisp() is called only by src\effects\nyquist\Nyquist.cpp but this source is not compiled if Nyquist is disabled:

https://github.com/audacity/audacity/blob/ea8163f94e58899844a173c679f5a267b4252d1c/src/CMakeLists.txt#L637

It is worth to apply here the same test on USE_NYQUIST macro, as it has been already done elsewhere.
This patch fixes and it closes issue #3137.