Separate classes TrackFocus and TrackPanelAx...
... The first is an attached object to the Project, the second is a wxWidgets
specific thing allocated with safenew and given to a wxWindow to destroy later,
and we don't know which will be destroyed first -- it varies with the operating
system.

So each needs to refer to the other with some kind of weak pointer.

But wxObjects should not be managed with std::weak_ptr, but instead, with
wxWeakRef.