FormatterContext is introduced...
FormatterContext is a class that makes AudacityProject to become an optional dependency of Formatters, while allowing to keep the old behavior as much as possible. 

FormatterContext is used to:
1) Filter the list of formatters. For example, if context has no project - beat formats have no sense.
2) Configuration for the formatter instance to use.

In case the FormatterContext has a valid project - it is the responsibility of the Formatter class to handle all the necessary changes to the format.