Identify NumericConverterRegistryItem differently...
... Don't prepend the group's identifier.  That is redundant for making unique
paths.

Don't encode the position within a group.  That would be fragile in case
reordering of the formats were wanted in some future release.  The identifier
of the format would not be unchanged across Audacity versions.

Use instead the internal string name that the BuiltinFormatString already has.

Is this very important? Not yet, because the identifiers are not written yet
into the config file, and no ordering hints yet refer to any of them.