issue #4568: Set selection boundary dialog format bug
Problem: The set/left selection boundary dialogs cannot display positions in any of the bar formats.

The NumericTextCtrls in the TimeDialog use the SampleRateContext, rather than the ProjectContext. The bar formats are not available in the SampleRateContext.

Fix: Change the NumericTextCtrls in the TimeDialog to use the ProjectContext (like they do in the Selection toolbar).
Note: TimeDialog::SetSampleRate() was removed as being no longer applicable, and it was not called by any existing code.