Play at speed slider: Problem with step sizes for keyboard users.
Problem:
Using the keyboard it can be not straightforward to get back to a speed of 1.00 after setting the slider to its minimum or maximum slider.

In void TranscriptionToolBar::Populate(), with the values set for options.Line() and options.Page(), and given how the step sizes are calculated in LWSlider::Increase() and LWSlider::Decrease(), the step sizes are 2.99/60, and 2.99/6.

Fix:
Change the values set for options.Line() and options.Page so that the step sizes are 3.00/60 and 3.00/6.
Note: After this fix, a user will still encounter the same issue if they set the slider to the minimum value, and then try to return to 1.00, but they can work around this by setting the slider to the maximum value and then returning to 1.00. In addition most users will probably be using speeds of > 1.00.