issue 4626: pasting label track to another project
Problem:
When pasting a label track to a different project using the "selected audio only" option, the label track can be time shifted.

Fix:
In DuplicateDiscardTrimmed(), in the call to trackCopy->SetOffset(), which restores the start times after the copy, pass track->GetStartTime(), rather than track->GetOffset().
For tracks other than a label track, GetOffset() and GetStartTime() return the same value. (So these tracks are not affected by this fix.)
However, for a label track, GetOffset() return 0.0.