WaveClip allows its channel count to be more than one...
... The number of channels does not change after construction (or
deserialization; consider that multi-part construction) of the WaveClip.

A copied WaveClip always has the same number of channels as its original.

Originally constructed (and not deserialized) WaveClips must have the number of
channels specified.  WaveTrack.cpp is the only file where those are constructed,
and for now that always passes 1.