Mixer ignores Track::GetChannel()...
... Note!  See how GetChannel() is really overridden by WaveTrack, as not a
simple accessor for the case that the track stores the MonoChannel value, but
dependent on pan and sometimes reporting LeftChannel or RightChannel instead.

The new logic is really equivalent although the proof is not quite trivial.

The IsMono() case will be visited where instead the IsLeft() or else branch of
the IsRight() case was visited.  But the effect is the same, assuming
numChannels is 1.