Pass leader tracks only into constructor of Mixer...
... This nicely eliminates some iterations elsewhere over channels instead of
tracks.

Note that in MixAndRender, right channel end time is now ignored so results may
change when it does not equal the left channel's end time.

Note that in Mixer::NeedsDither, identical rates in channels is assumed, and
again only the left channel's envelope is used, but now the override of
WidestSampleFormat checks for other channels.